Add parallel in vti diagnostics
